# docs/swagger.yaml
definitions:
...
  dao.UserCreate:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      user_from:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - phone_number
    - username
    type: object
  dao.UserPasswordUpdate:
    properties:
      id:
        type: integer
      password:
        type: string
      re_password:
        type: string
    required:
    - id
    - password
    - re_password
    type: object
  dao.UserUpdate:
    properties:
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      phone_number:
        type: string
    required:
    - id
    type: object
...
  service.MFAValidate:
    properties:
      code:
        type: string
      token:
        type: string
      username:
        type: string
    required:
    - code
    - token
    - username
    type: object
...
  service.RestPassword:
    properties:
      code:
        type: string
      password:
        type: string
      phone_number:
        type: string
      re_password:
        type: string
      username:
        type: string
    required:
    - code
    - password
    - phone_number
    - re_password
    - username
    type: object
...
  service.UserLogin:
    properties:
      ldap:
        type: boolean
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
...
  /api/v1/reset_password:
    post:
      description: 个人信息管理相关接口
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dao.UserPasswordUpdate'
      responses:
        "200":
          description: '{"code": 0, "msg": "更新成功"}'
          schema:
            type: string
      summary: 密码更新
      tags:
      - 个人信息管理
...
  /api/v1/sms/reset_password_code:
    post:
      consumes:
      - application/json
      description: 个人信息管理相关接口
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.RestPassword'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0, "msg": "校验码已发送..."}'
          schema:
            type: string
      summary: 获取验证码
      tags:
      - 个人信息管理
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dao.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0, "msg": "创建成功", "data": nil}'
          schema:
            type: string
      summary: 创建用户
      tags:
      - 用户管理
    put:
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dao.UserUpdate'
      responses:
        "200":
          description: '{"code": 0, "msg": "更新成功", "data": nil}'
          schema:
            type: string
      summary: 更新用户信息
      tags:
      - 用户管理
  /api/v1/user/{id}:
    delete:
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 0, "msg": "删除成功", "data": nil}'
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户管理
  /api/v1/user/avatarUpload:
    post:
      description: 个人信息管理相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 头像
        in: formData
        name: avatar
        required: true
        type: file
      responses:
        "200":
          description: '{"code": 0, "data": nil}'
          schema:
            type: string
      summary: 头像上传
      tags:
      - 个人信息管理
  /api/v1/user/info:
    get:
      description: 用户认证相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 0, "data": {}}'
          schema:
            type: string
      summary: 获取用户信息
      tags:
      - 用户认证
  /api/v1/user/list:
    get:
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 0, "data": []}'
          schema:
            type: string
      summary: 获取用户列表（下拉框）
      tags:
      - 用户管理
  /api/v1/user/mfa_auth:
    post:
      description: 用户认证相关接口
      parameters:
      - description: MFA认证信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.MFAValidate'
      responses:
        "200":
          description: '{"code": 0, "token": "用户令牌"}'
          schema:
            type: string
      summary: MFA认证
      tags:
      - 用户认证
  /api/v1/user/mfa_qrcode:
    get:
      description: 用户认证相关接口
      parameters:
      - description: 用户认证通过后的Token
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 0, "qrcode": ""}'
          schema:
            type: string
      summary: 获取MFA二维码
      tags:
      - 用户认证
  /api/v1/user/reset_mfa/{id}:
    put:
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 0, "msg": "重置成功", "data": nil}'
          schema:
            type: string
      summary: MFA重置
      tags:
      - 用户管理
  /api/v1/user/reset_password:
    put:
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dao.UserPasswordUpdate'
      responses:
        "200":
          description: '{"code": 0, "msg": "更新成功", "data": nil}'
          schema:
            type: string
      summary: 密码更新
      tags:
      - 用户管理
  /api/v1/user/sync/ad:
    post:
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 0, "msg": "同步成功"}'
          schema:
            type: string
      summary: AD域用户同步
      tags:
      - 用户管理
  /api/v1/users:
    get:
      description: 用户相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分页
        in: query
        name: page
        required: true
        type: integer
      - description: 分页大小
        in: query
        name: limit
        required: true
        type: integer
      - description: 用户姓名
        in: query
        name: name
        type: string
      responses:
        "200":
          description: '{"code": 0, "data": []}'
          schema:
            type: string
      summary: 获取查询的用户列表
      tags:
      - 用户管理
  /login:
    post:
      consumes:
      - application/json
      description: 用户认证相关接口
      parameters:
      - description: 用户名密码
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0, "token": "用户令牌"}'
          schema:
            type: string
      summary: 登录
      tags:
      - 用户认证
  /logout:
    post:
      description: 用户认证相关接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: '{"code": 0, "data": nil}'
          schema:
            type: string
      summary: 注销
      tags:
      - 用户认证
swagger: "2.0"